First Commits
Took out Privacy/Contact Footer and made an menu item. Contact us now
works on this project. Clicking on Forgot Password? goes to contact us
page. Re-arranged top menu so the left text is shifted all the way to
the left and the right menu items are shifted all the way to the right.
Added about menu to all types of menus(admins and users and login page).
Changed picture to the picture used on alumbook2. Changed name of
Internship menu item to Job & Internship Openings.