Worker/Admin adding account privileges and CSS fix
Made it so that the worker can only edit or delete user accounts and not
admins and other workers. Also, when creating a new account, the admin
can select between the three types(admin, worker, user) still but when
logged in as a worker, the worker does not have a selection since they
can only add users(created a hidden field that sets the type as user for
all new accounts that the worker creates). There was also a line going
through the screen on the edit accounts page for the admin, worker and
the user and fixed that by removing a class that was being added. I also
tested adding a avatar picture for a user and it works.