User Profile Info Work
Did a lot of work on User Profile since nothing was implemented
previously on alumbook1 for showing anything. I re-did tabs to match
alumbook2 and can show all profile info and edit as well. There are
still some things that need to be worked out(all the date stuff need to
be changed to datepickeers) and I still need to work on salary range,
some issues with the grad/undergrad degrees, the program selector and
the opt in stuff doesn't work yet(nothing saves if checked).